HIC: Split off HICObject and HICProperty
